Availability manager

+++ core/src/main/java/com/thundashop/core/availability/dto/Availability.java
@@ -16,5 +16,5 @@ public class Availability {
     private BigDecimal lowestPrice;
     private String currencyCode;
     private long totalNoOfRooms;
-    private Map<String, Long> totalNoOfRoomsByCategory;
+    private Map<String, Integer> totalNoOfRoomsByCategory;
 }
